Refactors Spring Security
