+++ hub/src/main/java/org/tapsensors/hub/config/security/EndpointSecurityConfigurer.java
@@ -4,14 +4,16 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.authentication.HttpStatusEntryPoint;
+import org.springframework.stereotype.Component;
 
-public abstract class BaseSecurityConfig {
-    protected void applyEndpointAuthorization(HttpSecurity httpSecurity) throws Exception {
+@Component
+public class EndpointSecurityConfigurer implements SecurityConfigurer{
+
+    @Override
+    public void configure(HttpSecurity httpSecurity) throws Exception {
         httpSecurity.exceptionHandling(e -> e.authenticationEntryPoint(new HttpStatusEntryPoint(HttpStatus.UNAUTHORIZED)));
-        httpSecurity.sessionManagement(s -> s.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
         httpSecurity.authorizeHttpRequests(http ->
                 http.requestMatchers("/readyz", "/livez").permitAll()
                         .anyRequest().authenticated());
     }
-    public abstract void applyOauth2ResourceServer(HttpSecurity httpSecurity) throws Exception;
 }
